requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
branca==0.6.0
celery==5.3.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2022.1
decorator==5.1.1
distlib==0.3.8
Django==4.2.5
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-filter==23.5
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
folium==0.14.0
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
geoip2==4.7.0
gunicorn==22.0.0
idna==2.10
inflection==0.5.1
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
maxminddb==2.4.0
multidict==6.0.4
numpy==1.25.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
ratelim==0.1.6
redis==5.0.0
requests==2.24.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.25.0
wcwidth==0.2.6
whitenoise==6.6.0
yarl==1.9.2
